Update solution model. Update solution handler
