+++ app/services/solution/solution.service.js
@@ -1,4 +1,11 @@
-export default {
+import Solution from '../../models/solution';
+import Task from '../../models/task';
+
+export default class SolutionService {
+  constructor(runner) {
+    this.runner = runner;
+  }
+
   saveSolutionRun({userId, taskId, solutionCode, result, Model}) {
     if (!(userId && taskId && solutionCode && result)) {
       return Promise.reject('required field is not specified');
@@ -28,4 +35,21 @@ export default {
     return new Model(data)
       .save();
   }
-};
+
+  async submitSolution(taskId, solutionCode, user) {
+    const task = await Task.findOne({id: taskId});
+    const runData = {
+      language: task.language,
+      code: solutionCode,
+      tests: task.test
+    };
+    const result = await this.runner.sendTask(runData);
+    return await this.saveSolutionRun({
+      userId: user._id,
+      taskId: task._id,
+      solutionCode,
+      result,
+      Model: Solution
+    });
+}
+}
